+++ deepxube/_train_cli.py
@@ -52,8 +52,9 @@ def parser_train(parser: ArgumentParser) -> None:
     update_group.add_argument('--procs', type=int, default=1, help="Number of processes to generate update data.")
     update_group.add_argument('--step_max', type=int, required=True, help="Maximum number of steps to take when generating problem instnaces.")
     update_group.add_argument('--up_itrs', type=int, default=100, help="Number of iterations to check for update.")
-    update_group.add_argument('--up_gen_itrs', type=int, default=100, help="Number of iterations for which to generate training data per update check.")
-    update_group.add_argument('--search_itrs', type=int, default=1000, help="Number of search iterations to take when generating data.")
+    update_group.add_argument('--up_gen_itrs', type=int, default=None, help="Number of iterations for which to generate training data per update check. "
+                                                                            "If None then defaults to up_itrs.")
+    update_group.add_argument('--search_itrs', type=int, default=1, help="Number of search iterations to take when generating data.")
     update_group.add_argument('--up_batch_size', type=int, default=100, help="Maximum number of problem instances to generate at a time. Lower if running out "
                                                                              "of memory.")
     update_group.add_argument('--up_nnet_batch_size', type=int, default=20000, help="Maximum number of inputs to give to any nnet at a time during update. "